refactor: step2 리뷰 피드백 적용
