+++ src/main/java/calculator/MinusOperator.java
@@ -1,4 +1,4 @@
-package step2;
+package calculator;
 
 public class MinusOperator implements Operator {
 	@Override
