+++ src/main/java/calculator/OperatorFactory.java
@@ -1,4 +1,4 @@
-package step2;
+package calculator;
 
 import java.util.HashMap;
 import java.util.Map;
